requirements.txt
# =============================================================================
# REQUIREMENTS - Prospection Pro Backend v5.1
# =============================================================================
# Compatible avec Railway, Render, Heroku et local
# =============================================================================

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# WebSocket
python-socketio==5.10.0
websockets==12.0

# Database - SQLAlchemy avec support SQLite et PostgreSQL
sqlalchemy==2.0.23
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Email
aiosmtplib==3.0.1
aioimaplib==1.0.1

# Async HTTP
aiohttp==3.9.1
httpx==0.25.2

# Data / Export Excel
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.2.0
pydantic==2.5.2

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1

# Production server
gunicorn==21.2.0
